+++ components/loan-list.js
@@ -1,6 +1,3 @@
-import Link from 'next/link'
-import { ArrowSmDownIcon, ArrowSmUpIcon } from '@heroicons/react/solid'
-import { UsersIcon } from '@heroicons/react/outline'
 import LoanCard from './loan-card'
 import { nanoErgsToErgs } from '../helper/erg-converter'
 
@@ -9,11 +6,11 @@ function classNames(...classes) {
 }
 
 export default function LoanList({ title, loanData }) {
-  const getFundingGoal = nanoErgs => {
+  const getFundingGoal = (nanoErgs) => {
     return 'Goal: ' + nanoErgsToErgs(nanoErgs) + ' Ergs'
   }
 
-  const getTitle = title => {
+  const getTitle = (title) => {
     if (title == null) {
       return 'Most Recent Loans'
     }
@@ -33,7 +30,7 @@ export default function LoanList({ title, loanData }) {
 
       <dl className="mt-5 grid grid-cols-1 gap-5 sm:grid-cols-2 lg:grid-cols-3">
         {loanData &&
-          loanData.loans.map(item => <LoanCard loan={item} key={item.id} />)}
+          loanData.loans.map((item) => <LoanCard loan={item} key={item.id} />)}
       </dl>
     </div>
   )
